backend/requirements.txt
autopep8==2.0.0
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.19
drf-yasg==1.21.4
